fix: remove BackstoryCampaignBehavior containing hardcoded lord ids Bannerlord randomly creates new child characters for some factions based on an existing lord. Sometimes characters with id lord_1_1 or lord_1_7 are selected.
Their body properties have a custom hair_tag value which makes the game crash.
So we could just remove the hair_tag and it will fix the crash.

However, these lords are defined in our xml due to being hardcoded in the BackstoryCampaignBehavior object.
So, we should remove these useless lords and remove that campaign behaviour.
It basically applies grudges between native lords. It should not apply to our mod as we have our own lords.

DellarteDellaGuerraTeam/DellarteDellaGuerraMap#14